Move alias map builder to its own file
