+++ external-crates/move/crates/move-compiler/src/expansion/aliases.rs
@@ -3,60 +3,15 @@
 // SPDX-License-Identifier: Apache-2.0
 
 use crate::{
-    expansion::ast::{self as E, ModuleIdent},
-    expansion::translate::ModuleMemberKind,
-    parser::ast::{self as P},
+    expansion::alias_map_builder::*,
     shared::{unique_map::UniqueMap, unique_set::UniqueSet, *},
 };
-use move_ir_types::location::*;
 use std::{collections::BTreeSet, fmt};
 
 #[derive(Clone, Debug)]
 pub struct AliasSet {
     pub modules: UniqueSet<Name>,
     pub members: UniqueSet<Name>,
-}
-
-pub type MemberName = (ModuleIdent, Name, ModuleMemberKind);
-
-#[derive(Clone)]
-pub enum AliasMapBuilder {
-    Legacy {
-        modules: UniqueMap<Name, (ModuleIdent, /* is_implicit */ bool)>,
-        members: UniqueMap<Name, (MemberName, /* is_implicit */ bool)>,
-    },
-    Namespaced {
-        leading_access: UniqueMap<Name, (LeadingAccessEntry, /* is_implicit */ bool)>,
-        module_members: UniqueMap<Name, (MemberEntry, /* is_implicit */ bool)>,
-    },
-}
-
-#[derive(Clone, Copy, PartialEq, Eq, PartialOrd, Ord)]
-pub enum AliasEntry {
-    Address(Name, NumericalAddress),
-    Module(Name, ModuleIdent),
-    Member(Name, ModuleIdent, Name),
-    TypeParam(Name),
-}
-
-#[derive(Clone, Copy)]
-pub enum LeadingAccessEntry {
-    Address(NumericalAddress),
-    Module(ModuleIdent),
-    Member(ModuleIdent, Name),
-    TypeParam,
-}
-
-#[derive(Clone, Copy)]
-pub enum MemberEntry {
-    Member(ModuleIdent, Name),
-    TypeParam,
-}
-
-#[derive(Clone, Copy)]
-pub enum NameSpace {
-    LeadingAccess,
-    ModuleMembers,
 }
 
 pub struct AliasMap {
@@ -70,20 +25,6 @@ pub struct AliasMap {
     previous: Option<Box<AliasMap>>,
 }
 
-pub struct ParserExplicitUseFun {
-    pub loc: Loc,
-    pub attributes: E::Attributes,
-    pub is_public: Option<Loc>,
-    pub function: Box<P::NameAccessChain>,
-    pub ty: Box<P::NameAccessChain>,
-    pub method: Name,
-}
-
-pub struct UseFunsBuilder {
-    pub explicit: Vec<ParserExplicitUseFun>,
-    pub implicit: UniqueMap<Name, E::ImplicitUseFunCandidate>,
-}
-
 impl AliasSet {
     pub fn new() -> Self {
         Self {
@@ -95,207 +36,6 @@ impl AliasSet {
     #[allow(unused)]
     pub fn is_empty(&self) -> bool {
         self.modules.is_empty() && self.members.is_empty()
-    }
-}
-
-/// Remove a duplicate element in the map, returning its location as an error if it exists
-fn remove_dup<K: TName, V>(map: &mut UniqueMap<K, V>, alias: &K) -> Result<(), K::Loc> {
-    let loc = map.get_loc(alias).copied();
-    match map.remove(alias) {
-        None => Ok(()),
-        Some(_) => Err(loc.unwrap()),
-    }
-}
-
-impl AliasMapBuilder {
-    /// Create a new AliasMapBuilder for legacy
-    pub fn legacy() -> Self {
-        Self::Legacy {
-            modules: UniqueMap::new(),
-            members: UniqueMap::new(),
-        }
-    }
-
-    pub fn namespaced() -> Self {
-        Self::Namespaced {
-            leading_access: UniqueMap::new(),
-            module_members: UniqueMap::new(),
-        }
-    }
-
-    pub fn is_empty(&self) -> bool {
-        match self {
-            Self::Legacy { modules, members } => modules.is_empty() && members.is_empty(),
-            Self::Namespaced {
-                leading_access,
-                module_members,
-            } => leading_access.is_empty() && module_members.is_empty(),
-        }
-    }
-
-    fn remove_module_alias(&mut self, alias: &Name) -> Result<(), Loc> {
-        match self {
-            Self::Legacy { modules, .. } => remove_dup(modules, alias),
-            Self::Namespaced { leading_access, .. } => remove_dup(leading_access, alias),
-        }
-    }
-
-    fn remove_member_alias(&mut self, alias: &Name, kind: ModuleMemberKind) -> Result<(), Loc> {
-        match (self, kind) {
-            (AliasMapBuilder::Legacy { members, .. }, _) => remove_dup(members, alias),
-            // constants and functions are not in the leading access namespace
-            (
-                AliasMapBuilder::Namespaced {
-                    leading_access: _,
-                    module_members,
-                },
-                ModuleMemberKind::Constant | ModuleMemberKind::Function | ModuleMemberKind::Schema,
-            ) => remove_dup(module_members, alias),
-            // structs are in the leading access namespace in addition to the module members
-            // namespace
-            (
-                AliasMapBuilder::Namespaced {
-                    leading_access,
-                    module_members,
-                },
-                ModuleMemberKind::Struct,
-            ) => {
-                let r1 = remove_dup(module_members, alias);
-                let r2 = remove_dup(leading_access, alias);
-                r1.and(r2)
-            }
-        }
-    }
-
-    fn remove_address_alias(&mut self, alias: &Name) -> Result<(), Loc> {
-        match self {
-            Self::Legacy { .. } => Ok(()),
-            Self::Namespaced { leading_access, .. } => remove_dup(leading_access, alias),
-        }
-    }
-
-    /// Adds a module alias to the map.
-    /// Errors if one already bound for that alias
-    fn add_module_alias_(
-        &mut self,
-        alias: Name,
-        ident: ModuleIdent,
-        is_implicit: bool,
-    ) -> Result<(), Loc> {
-        let result = self.remove_module_alias(&alias);
-        match self {
-            Self::Legacy { modules, .. } => modules.add(alias, (ident, is_implicit)).unwrap(),
-            Self::Namespaced { leading_access, .. } => {
-                let entry = (LeadingAccessEntry::Module(ident), is_implicit);
-                leading_access.add(alias, entry).unwrap()
-            }
-        }
-        result
-    }
-
-    fn add_member_alias_(
-        &mut self,
-        alias: Name,
-        ident: ModuleIdent,
-        member: Name,
-        kind: ModuleMemberKind,
-        is_implicit: bool,
-    ) -> Result<(), Loc> {
-        let result = self.remove_member_alias(&alias, kind);
-        match (self, kind) {
-            (AliasMapBuilder::Legacy { members, .. }, _) => members
-                .add(alias, ((ident, member, kind), is_implicit))
-                .unwrap(),
-            // constants and functions are not in the leading access namespace
-            (
-                AliasMapBuilder::Namespaced {
-                    leading_access: _,
-                    module_members,
-                },
-                ModuleMemberKind::Constant | ModuleMemberKind::Function | ModuleMemberKind::Schema,
-            ) => {
-                let entry = (MemberEntry::Member(ident, member), is_implicit);
-                module_members.add(alias, entry).unwrap();
-            }
-            // structs are in the leading access namespace in addition to the module members
-            // namespace
-            (
-                AliasMapBuilder::Namespaced {
-                    leading_access,
-                    module_members,
-                },
-                ModuleMemberKind::Struct,
-            ) => {
-                let member_entry = (MemberEntry::Member(ident, member), is_implicit);
-                module_members.add(alias, member_entry).unwrap();
-                let leading_access_entry = (LeadingAccessEntry::Member(ident, member), is_implicit);
-                leading_access.add(alias, leading_access_entry).unwrap();
-            }
-        }
-        result
-    }
-
-    fn add_address_alias_(
-        &mut self,
-        alias: Name,
-        address: NumericalAddress,
-        is_implicit: bool,
-    ) -> Result<(), Loc> {
-        let result = self.remove_address_alias(&alias);
-        match self {
-            Self::Legacy { .. } => (),
-            Self::Namespaced { leading_access, .. } => {
-                let entry = (LeadingAccessEntry::Address(address), is_implicit);
-                leading_access.add(alias, entry).unwrap()
-            }
-        }
-        result
-    }
-
-    /// Adds a module alias to the map.
-    /// Errors if one already bound for that alias
-    pub fn add_module_alias(&mut self, alias: Name, ident: ModuleIdent) -> Result<(), Loc> {
-        self.add_module_alias_(alias, ident, /* is_implicit */ false)
-    }
-
-    /// Adds a member alias to the map.
-    /// Errors if one already bound for that alias
-    pub fn add_member_alias(
-        &mut self,
-        alias: Name,
-        ident: ModuleIdent,
-        member: Name,
-        kind: ModuleMemberKind,
-    ) -> Result<(), Loc> {
-        self.add_member_alias_(alias, ident, member, kind, /* is_implicit */ false)
-    }
-
-    /// Adds an address alias to the map.
-    /// Errors if one already bound for that alias
-    pub fn add_address_alias(&mut self, alias: Name, address: NumericalAddress) -> Result<(), Loc> {
-        self.add_address_alias_(alias, address, /* is_implicit */ false)
-    }
-
-    /// Same as `add_module_alias` but it does not update the scope, and as such it will not be
-    /// reported as unused
-    pub fn add_implicit_module_alias(
-        &mut self,
-        alias: Name,
-        ident: ModuleIdent,
-    ) -> Result<(), Loc> {
-        self.add_module_alias_(alias, ident, /* is_implicit */ true)
-    }
-
-    /// Same as `add_member_alias` but it does not update the scope, and as such it will not be
-    /// reported as unused
-    pub fn add_implicit_member_alias(
-        &mut self,
-        alias: Name,
-        ident: ModuleIdent,
-        member: Name,
-        kind: ModuleMemberKind,
-    ) -> Result<(), Loc> {
-        self.add_member_alias_(alias, ident, member, kind, /* is_implicit */ true)
     }
 }
 
@@ -447,107 +187,9 @@ impl AliasMap {
     }
 }
 
-impl From<(Name, NumericalAddress)> for AliasEntry {
-    fn from((name, addr): (Name, NumericalAddress)) -> Self {
-        AliasEntry::Address(name, addr)
-    }
-}
-
-impl From<(Name, LeadingAccessEntry)> for AliasEntry {
-    fn from((name, entry): (Name, LeadingAccessEntry)) -> Self {
-        match entry {
-            LeadingAccessEntry::Address(addr) => AliasEntry::Address(name, addr),
-            LeadingAccessEntry::Module(mident) => AliasEntry::Module(name, mident),
-            LeadingAccessEntry::Member(mident, member) => AliasEntry::Member(name, mident, member),
-            LeadingAccessEntry::TypeParam => AliasEntry::TypeParam(name),
-        }
-    }
-}
-
-impl From<(Name, MemberEntry)> for AliasEntry {
-    fn from((name, entry): (Name, MemberEntry)) -> Self {
-        match entry {
-            MemberEntry::Member(mident, member) => AliasEntry::Member(name, mident, member),
-            MemberEntry::TypeParam => AliasEntry::TypeParam(name),
-        }
-    }
-}
-
-impl UseFunsBuilder {
-    pub fn new() -> Self {
-        Self {
-            explicit: vec![],
-            implicit: UniqueMap::new(),
-        }
-    }
-}
-
 //**************************************************************************************************
-// Display
+// Debug
 //**************************************************************************************************
-
-impl fmt::Debug for AliasEntry {
-    fn fmt(&self, f: &mut fmt::Formatter) -> std::fmt::Result {
-        match self {
-            AliasEntry::Module(alias, mident) => write!(f, "({alias}, m#{mident})"),
-            AliasEntry::Address(alias, addr) => write!(f, "({alias}, @{addr})"),
-            AliasEntry::Member(alias, mident, name) => write!(f, "({alias},{mident}::{name})"),
-            AliasEntry::TypeParam(alias) => write!(f, "({alias},[tparam])"),
-        }
-    }
-}
-
-impl fmt::Debug for LeadingAccessEntry {
-    fn fmt(&self, f: &mut fmt::Formatter) -> std::fmt::Result {
-        match self {
-            LeadingAccessEntry::Module(mident) => write!(f, "m#{mident}"),
-            LeadingAccessEntry::Address(addr) => write!(f, "@{addr}"),
-            LeadingAccessEntry::Member(mident, name) => write!(f, "{mident}::{name}"),
-            LeadingAccessEntry::TypeParam => write!(f, "[tparam]"),
-        }
-    }
-}
-
-impl fmt::Debug for MemberEntry {
-    fn fmt(&self, f: &mut fmt::Formatter) -> std::fmt::Result {
-        match self {
-            MemberEntry::Member(mident, name) => write!(f, "{mident}::{name}"),
-            MemberEntry::TypeParam => write!(f, "[tparam]"),
-        }
-    }
-}
-
-impl fmt::Debug for AliasMapBuilder {
-    fn fmt(&self, f: &mut fmt::Formatter) -> std::fmt::Result {
-        match self {
-            AliasMapBuilder::Legacy { modules, members } => {
-                writeln!(f, "AliasMapBuilder::Legacy(\n  modules: [")?;
-                for (_, key, (target, is_implicit)) in modules {
-                    writeln!(f, "    {key} => {target:?} <{is_implicit}>,")?;
-                }
-                writeln!(f, "  ],\n  members: [")?;
-                for (_, key, (target, is_implicit)) in members {
-                    writeln!(f, "    {key} => {target:?} <{is_implicit}>,")?;
-                }
-                writeln!(f, "])")
-            }
-            AliasMapBuilder::Namespaced {
-                leading_access,
-                module_members,
-            } => {
-                writeln!(f, "AliasMapBuilder::Legacy(\n  leading_access: [")?;
-                for (_, key, (target, is_implicit)) in leading_access {
-                    writeln!(f, "    {key} => {target:?} <{is_implicit}>,")?;
-                }
-                writeln!(f, "  ],\n  module_members: [")?;
-                for (_, key, (target, is_implicit)) in module_members {
-                    writeln!(f, "    {key} => {target:?} <{is_implicit}>,")?;
-                }
-                writeln!(f, "])")
-            }
-        }
-    }
-}
 
 impl fmt::Debug for AliasMap {
     fn fmt(&self, f: &mut fmt::Formatter) -> std::fmt::Result {
