+++ external-crates/move/crates/move-compiler/src/expansion/mod.rs
@@ -2,6 +2,7 @@
 // Copyright (c) The Move Contributors
 // SPDX-License-Identifier: Apache-2.0
 
+mod alias_map_builder;
 mod aliases;
 pub mod ast;
 mod byte_string;
